Cleaner match that should work

+++ include/REL/m401Dll.h
@@ -72,7 +72,7 @@ void fn_2_1041C(void);
 void fn_2_10710(void);
 
 void fn_2_1079C(void);
-omObjData *fn_2_10A88(Vec *arg0, s32 arg1);
+omObjData *fn_2_10A88(Vec *arg0, s16 arg1);
 
 omObjData *fn_2_141B0(void);
 omObjData *fn_2_14640(Process *objman);
